Update detectLineVideo.py
Detection of Road Lane Lines

Instructions:

A Live Lane-Line Detection Systems built-in Python language is another Data Science project idea for beginners. A human driver receives lane detecting instruction from lines placed on the road in this project. The lines placed on the roads indicate where the lanes are located for human driving. It also refers to the vehicle’s steering direction. This application is crucial for the development of self-driving cars. This application for the Data Science Project is critical for the development of self-driving cars.